+++ app.ts
@@ -1,6 +1,14 @@
 /// <reference path="./wx.d.ts" />
 
 //在微信小程序的app.ts中这样用
+const options = {
+  duration: 10000,
+  sampleRate: 44100,
+  numberOfChannels: 1,
+  encodeBitRate: 192000,
+  format: "aac",
+  frameSize: 50
+}
 Page({
   data:{},
   a(){
@@ -17,5 +25,16 @@ Page({
     // wx.onSocketOpen(res => console.log(res.header))
     wx.onSocketMessage(res => console.log(res.data));
     wx.onSocketClose(res => console.log(res))
+    const recorderManager = wx.getRecorderManager()
+    recorderManager.onStop((res) => console.log(res.tempFilePath))
+    recorderManager.onPause(() => console.log('dadadadad'))
+    recorderManager.start(options)
+    wx.getImageInfo({
+      src: 'images/a.jpg',
+      success: function (res) {
+        console.log(res.orientation)
+        console.log(res.height)
+      }
+    })
   }
 })